Fix synchronisation issue with TestCache.
When bridgestrap tests several bridges in parallel, it spawns several go
routines.  When all these go routines access our cache simultaneously,
we're running into sync issues because more than one go routine is
reading from or writing to the cache.

This commit adds a locking mechanism to fix this issue, and a unit test
to demonstrate that the fix works.