Add -cache-timeout command line switch.
So far, our cache timeout was hard-coded to be one week.  Now that
bridgestrap is more efficient, we can afford to test resources more
often, so this patch changes the default timeout to 24 hours.  The patch
also makes this timeout configurable through the new -cache-timeout
switch.

While at it, the patch also refactored our cache mechanism a bit and
renamed -timeout to -test-timeout.

This patch fixes tpo/anti-censorship/bridgestrap#8