Make metrics more idiomatic to Prometheus.
Our metrics aren't implemented in the way Prometheus recommends.  This
patch makes use of labels and renames our metrics to change that.  For
example, instead of having two metrics called cache_hits and
cache_misses, we now have a single metric called cache_total and it
accepts a "type" label that can take on the values "hit" or "miss".

This patch fixes tpo/anti-censorship/bridgestrap#10.