Try waking up tor before giving it new bridges.
By default, tor goes to sleep after 24 hours of no user activity.
Bridgestrap's control port interaction does not count as user activity,
which means that tor effectively stops working after 24 hours.  To work
around this, we're sending tor a SIGNAL ACTIVE, which wakes up tor in
case it went to sleep.  This is a cheap thing to do, so we might as well
do it before each SETCONF.

This patch fixes tpo/anti-censorship/bridgestrap#12.