Fix bug where test cache timeout was ignored.
When bridgestrap read its cache from disk, it wouldn't only read the
cache entries but also the cache timeout, overriding whatever we
provided over the command line.  In other words, if bridgestrap saved a
cache whose timeout was set to 24h, and we later restart bridgestrap
with a given cache timeout of 18h, it would use 24h because that's what
it read from disk.

This patch fixes this bug in two ways: 1) we no longer write the cache
timeout to disk and 2) we set the cache timeout after reading the cached
entries from disk.